requirements.txt
boto3~=1.15
termcolor~=1.1
pyyaml~=5.1
docker~=3.7
